backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.5.2
motor==3.6.0
pymongo==4.9.1
httpx==0.27.2
